PHP-117- Edited JsonShoppingListRepo and its test

+++ src/Persistance/Json/JsonShoppingListRepository.php
@@ -100,7 +100,8 @@ class JsonShoppingListRepository implements ShoppingListRepositoryInterface
     {
         $filename = $this->storeAs($slug);
 
-        $this->files->unlink($filename);
+        if ($this->files->exists($filename)) {
+            $this->files->unlink($filename);
+        }
     }
-
 }
